Convert Timer() thread to ScheduledThreadPoolExecutor() service
