+++ samigo/samigo-app/src/java/org/sakaiproject/tool/assessment/ui/queue/delivery/TimedAssessmentQueue.java
@@ -19,92 +19,101 @@
  *
  **********************************************************************************/
 
-
 package org.sakaiproject.tool.assessment.ui.queue.delivery;
-import java.util.Collection;
-import java.util.Iterator;
-import java.util.Timer;
-import java.util.concurrent.ConcurrentHashMap;
 
+import java.util.Collection;
+import java.util.HashMap;
+import java.util.Iterator;
+import java.util.concurrent.ConcurrentHashMap;
+import java.util.concurrent.ScheduledThreadPoolExecutor;
+import java.util.concurrent.ScheduledFuture;
+import java.util.concurrent.Executors;
+import java.util.concurrent.TimeUnit;
+import java.util.ResourceBundle;
+import org.sakaiproject.component.cover.ServerConfigurationService;
+import org.sakaiproject.event.cover.EventTrackingService;
+import org.sakaiproject.event.cover.NotificationService;
+import org.sakaiproject.samigo.util.SamigoConstants;
+import org.sakaiproject.tool.api.Session;
+import org.sakaiproject.tool.assessment.ui.model.delivery.TimedAssessmentGradingModel;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.sakaiproject.tool.assessment.ui.model.delivery.TimedAssessmentGradingModel;
+
 
 /**
- * <p>Title: </p>
- * <p>Description: </p>
- * <p>Copyright: Copyright (c) 2004</p>
- * <p>Company: </p>
- * @author not attributable
- * @version 1.0
+ * <p>Title: TimedAssessmentQueue</p>
+ * <p>Description: A queue of assessment timers executed in a pool of threads</p>
  */
 
 public class TimedAssessmentQueue { 
-  private static TimedAssessmentQueue instance; 
+
   private ConcurrentHashMap queue;
-    // private SubmitTimedAssessmentThread thread;
-  private Timer timer;
-
+  private HashMap<Long, ScheduledFuture<?>> tasks;
+  private ScheduledThreadPoolExecutor threadPool;
   private static Logger log = LoggerFactory.getLogger(TimedAssessmentQueue.class);
-  private TimedAssessmentQueue() { 
-    queue = new ConcurrentHashMap ();
-  } 
 
-  public static synchronized TimedAssessmentQueue getInstance() { 
-    if (instance==null) { 
-      instance = new TimedAssessmentQueue(); 
-    } 
-    return instance; 
-  } 
-  
-  public void add(TimedAssessmentGradingModel timedAG){
-	  synchronized (queue) {
-		  queue.put(timedAG.getAssessmentGradingId(), timedAG);
-		  log.debug("***1. TimedAssessmentQueue.add, before schedule timer="+timer);
-		  scheduleTask();
-		  log.debug("***2. TimedAssessmentQueue.add, after schedule timer="+timer);
-	  }
+  private static class Loader {
+    static final TimedAssessmentQueue INSTANCE = new TimedAssessmentQueue();
   }
 
-  private void scheduleTask(){
-    if (timer == null){
-      timer  = new Timer();
-      timer.scheduleAtFixedRate(new SubmitTimedAssessmentThread(), 0, 3*1000); //delay=0s, every 3s 
+  private TimedAssessmentQueue() { 
+    queue = new ConcurrentHashMap();
+    tasks = new HashMap();
+    // Get any custom thread count, or default to 4
+    int threads = ServerConfigurationService.getInt("samigo.timerThreadCount", 4);
+    log.info( "SAMIGO_TIMED_ASSESSMENT:QUEUE:INIT: THREADS:" + threads + " (Set property 'samigo.timerThreadCount' to adjust.)");
+    threadPool = new ScheduledThreadPoolExecutor(threads);
+    // Set the default removal policy so we don't leak memory with finished tasks
+    threadPool.setRemoveOnCancelPolicy(true);
+  } 
+
+
+  // Get the instance of this class
+  public static TimedAssessmentQueue getInstance() { 
+    return Loader.INSTANCE;
+  } 
+
+  
+  // Add a timed assessment to the queue
+  public void add(TimedAssessmentGradingModel timedAG){
+          // Add grading data to the queue
+          synchronized (queue) {
+                  queue.put(timedAG.getAssessmentGradingId(),timedAG);
+          }
+          try {
+              // Repeat the task until stopped or an exception is thrown
+              // Store the resulting ScheduledFuture for the task
+              // delay=0s, every 3s 
+              tasks.put(timedAG.getAssessmentGradingId(), threadPool.scheduleAtFixedRate(new TimedAssessmentRunnable(timedAG.getAssessmentGradingId()), 0, 3000, TimeUnit.MILLISECONDS)); 
+              log.info( "SAMIGO_TIMED_ASSESSMENT:QUEUE:NEWTIMER:SUCCESS ID:" + timedAG.getAssessmentGradingId());
+          } catch (Exception ex) {
+              log.error("SAMIGO_TIMED_ASSESSMENT:QUEUE:NEWTIMER:FAILED ID:" + timedAG.getAssessmentGradingId() + " Exception:" + ex);
+          }
+  }
+
+
+  // Remove a timed assessment from the queue
+  public void remove(TimedAssessmentGradingModel timedAG){
+          remove(timedAG.getAssessmentGradingId());
+  }
+
+
+  // Remove a timed assessment from the queue
+  public void remove(long timedAG){
+    log.info("SAMIGO_TIMED_ASSESSMENT:QUEUE:REMOVE ID:" + timedAG);
+    // Stop the task and remove it
+    tasks.get(timedAG).cancel(true);
+    tasks.remove(timedAG);
+    // Remove the grading data from the queue
+    synchronized (queue) {
+      queue.remove(timedAG);
     }
   }
 
-  public void remove(TimedAssessmentGradingModel timedAG){
-	  synchronized (queue) {
-		  queue.remove(timedAG.getAssessmentGradingId());
-		  if (isEmpty()){
-			  log.debug("*** before destroy, timer="+timer);
-			  timer.cancel();
-			  timer = null;
-			  log.debug("*** after destroy, timer="+timer);
-		  }
-	  }
-  }
-  
-  public TimedAssessmentGradingModel get(Long assessmentGradingId){
+
+  // Get assessment data from the queue
+  public TimedAssessmentGradingModel get(long assessmentGradingId){
     return (TimedAssessmentGradingModel)queue.get(assessmentGradingId);
   }
 
-  public void emptyQueue(){
-    queue.clear();
-  }
-
-  public boolean isEmpty(){
-    return queue.isEmpty();
-  }
-
-  public int size(){
-    return queue.size();
-  }
-
-  public Iterator iterator(){
-    Collection c = queue.values(); 
-    return c.iterator();
-  }
-
-} 
-
+}
